pyproject.toml
[project]
name = "weather-agent-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.107.1",
    "openai-agents>=0.3.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
]
